split stationary and onTheFly tests
