+++ theRobot/src/test/java/frc/robot/control/ShootOnTheFlyTest.java
@@ -34,138 +34,6 @@ class ShootOnTheFlyTest {
     subsystemCollection.setCameraSubsystem(null);
     drivetrain = new DrivetrainSubsystem(subsystemCollection);
     shooterAimer = new ShooterAimer(drivetrain, subsystemCollection);
-  }
-
-  @AfterEach // this method will run after each test
-  void shutdown() throws Exception {
-    // need to close drivetrain and shooterAimer here...
-  }
-
-  /*
-   * Static shooting tests
-   */
-
-  @Test
-  void hoodExtensionValidValue() {
-    // known lookup table value
-    double input = 3.4408;
-    double output = 0.45;
-    double hoodPosition = shooterAimer.hoodExtensionForDistance(input);
-    assertEquals(
-        output, hoodPosition, DELTA);
-  }
-
-  @Test
-  void hoodExtensionInterpolatedValue() {
-    // interpolated lookup table value
-    // 3.4408, 0.45 },
-    // { 4.7448, 0.637
-    double input = 4.0;
-    double numerator = (0.45 * (4.7448 - input)) + (0.637 * (input - 3.4408));
-    double denominator = (4.7448 - 3.4408);
-    double output = numerator / denominator;
-    double hoodPosition = shooterAimer.hoodExtensionForDistance(input);
-    assertEquals(
-        output, hoodPosition, DELTA);
-  }
-
-  @Test
-  void hoodExtensionClampBelowMin() {
-    // clamp to min when input below min
-    double input = 0.5; // below min
-    double output = Constants.hoodMinPositionRotations;
-    double hoodPosition = shooterAimer.hoodExtensionForDistance(input);
-    assertEquals(
-        output, hoodPosition, DELTA);
-  }
-
-  @Test
-  void hoodExtensionClampAboveMax() {
-    // clamp to max when input above max
-    double input = 9.0; // below min
-    double output = Constants.hoodMaxPositionRotations;
-    double hoodPosition = shooterAimer.hoodExtensionForDistance(input);
-    assertEquals(
-        output, hoodPosition, DELTA);
-  }
-
-  @Test
-  void shooterRPMValidValue() {
-    // known lookup table value
-    double input = 1.3037;
-    double output = 3000;
-    double hoodPosition = shooterAimer.shooterRpmForDistance(input);
-    assertEquals(
-        output, hoodPosition, DELTA);
-  }
-
-  @Test
-  void shooterRPMInterpolatedValue() {
-    // interpolated lookup table value
-    // { 1.3037, 3000 },
-    // { 4.7448, 4000 },
-    double input = 4.0;
-    double numerator = (3000 * (4.7448 - input)) + (4000 * (input - 1.3037));
-    double denominator = (4.7448 - 1.3037);
-    double output = numerator / denominator;
-    double hoodPosition = shooterAimer.shooterRpmForDistance(input);
-    assertEquals(
-        output, hoodPosition, DELTA);
-  }
-
-  @Test
-  void shooterRPMClampBelowMin() {
-    // clamp to min when input below min
-    double input = 0.5; // below min
-    double output = 2912;
-    double hoodPosition = shooterAimer.shooterRpmForDistance(input);
-    assertEquals(
-        output, hoodPosition, DELTA);
-  }
-
-  @Test
-  void ShooterRPMClampAboveMax() {
-    // clamp to max when input above max
-    double input = 9.0; // below min
-    double output = 6500;
-    double hoodPosition = shooterAimer.shooterRpmForDistance(input);
-    assertEquals(
-        output, hoodPosition, DELTA);
-  }
-
-  @Test
-  void RobotYawTargetDeadOn() {
-    // robot yaw should be 180 when dead on to target
-    drivetrain
-        .setRobotPosition(new Pose2d(Constants.blueHubPosition.plus(new Translation2d(1.0, 0.0)), new Rotation2d(0.0)));
-    assertEquals(Rotation2d.fromDegrees(180.0).plus(Constants.shooterYawOffset).getDegrees(), shooterAimer.getYawToFaceTarget(Constants.blueHubPosition).getDegrees());
-  }
-
-  @Test
-  void RobotYawTarget45degrees() {
-    // robot yaw should be -135 (-180+45 - shooterYawOffset) degrees when offset equal amounts in x/y
-    drivetrain
-        .setRobotPosition(new Pose2d(Constants.blueHubPosition.plus(new Translation2d(1.0, 1.0)), new Rotation2d(0.0)));
-    assertEquals(Rotation2d.fromDegrees(-180.0 + 45.0).plus(Constants.shooterYawOffset).getDegrees(), shooterAimer.getYawToFaceTarget(Constants.blueHubPosition).getDegrees());
-  }
-
-  @Test
-  void RobotYawTarget90degrees() {
-    // robot yaw should be -90 when to the side of target
-    drivetrain
-        .setRobotPosition(new Pose2d(Constants.blueHubPosition.plus(new Translation2d(0.0, 1.0)), new Rotation2d(0.0)));
-    assertEquals(Rotation2d.fromDegrees(-90.0).plus(Constants.shooterYawOffset).getDegrees(), shooterAimer.getYawToFaceTarget(Constants.blueHubPosition).getDegrees());
-  }
-
-  /*
-   * Shoot on the fly tests
-   */
-  
-  @Test
-  void RobotYawTargetDeadOnFadeShot() {
-    // robot yaw should be 180 when dead on to target
-    // target should move farther when driving away
-
     // enable driver station and set alliance
     DriverStationSim.setAllianceStationId(AllianceStationID.Blue1);
     DriverStationSim.setEnabled(true);
@@ -175,21 +43,29 @@ class ShootOnTheFlyTest {
 
     // set target
     shooterAimer.setDesiredTarget(shooterAimer.getHubPositionFromAlliance());
+  }
+
+  @AfterEach // this method will run after each test
+  void shutdown() throws Exception {
+    // need to close drivetrain and shooterAimer here...
+  }
+
+  @Test
+  void RobotYawTargetDeadOnFadeShot() {
+    // robot yaw should be 180 when dead on to target
+    // target should move farther when driving away
     
     // setup drivetrain
     // start 1m from target in x
     // have to set the drivetrain x/y first, and then apply the auto yaw (since the PID that would set it doesn't work in sim)
     drivetrain
         .setRobotPosition(new Pose2d(Constants.blueHubPosition.plus(new Translation2d(1.0, 0.0)), new Rotation2d(0.0)));
-    System.out.println("auto yaw rotational velocity " + shooterAimer.computeAutoYawVelocityRadiansPerSecond());
     drivetrain
         .setRobotPosition(new Pose2d(Constants.blueHubPosition.plus(new Translation2d(1.0, 0.0)), shooterAimer.getYawToFaceTarget(Constants.blueHubPosition)));
-    System.out.println("auto yaw rotational velocity after" + shooterAimer.computeAutoYawVelocityRadiansPerSecond());
-    // drive 1mps away from target in x
+    // drive 1 mps away from target in x
     double vx = -1.0;
     double vy = 0.0;
     drivetrain.driveFieldCentricShooting(new ChassisSpeeds(vx, vy, 0.0));
-    System.out.println("field centric chassis speeds " + drivetrain.getChassisSpeedsFieldCentric());
 
     // compute expected target
     Translation2d expectedTarget = Constants.blueHubPosition.plus(
@@ -199,6 +75,24 @@ class ShootOnTheFlyTest {
     System.out.println("expected target " + expectedTarget);
     System.out.println("actual target " + shooterAimer.computePredictedTarget());
     assertTrue(expectedTarget.equals(shooterAimer.computePredictedTarget()));
+  }
+
+  @Test
+  void RobotYawRotationalVelocityMatters() {
+    // setup drivetrain
+    // start 1m from target in x
+    // have to set the drivetrain x/y first, and then apply the auto yaw (since the PID that would set it doesn't work in sim)
+    drivetrain
+        .setRobotPosition(new Pose2d(Constants.blueHubPosition.plus(new Translation2d(1.0, 0.0)), new Rotation2d(0.0)));
+    // make position 5 degrees off and see what the PID speed is
+    drivetrain
+        .setRobotPosition(new Pose2d(Constants.blueHubPosition.plus(new Translation2d(1.0, 0.0)), 
+        shooterAimer.getYawToFaceTarget(Constants.blueHubPosition).plus(Rotation2d.fromDegrees(5))));
+    double yawVelocityRadiansPerSecond = shooterAimer.computeAutoYawVelocityRadiansPerSecond();
+    System.out.println("auto yaw rotational velocity " + yawVelocityRadiansPerSecond);
+    double translationalVelocity = yawVelocityRadiansPerSecond * Constants.shooterOffsetFromCenterOfRobot.getNorm();
+    System.out.println("auto yaw translations velocity (mps)" + translationalVelocity);
+
   }
 
 }
